[QRChecker] Cog model restructure
- Restructure `QRChecker` according to SFUAnime/Ren#608.
  No functional changes.
- Move constants to `constants.py`.
- Remove redundant type hints (e.g., those fully inferable from the
  return types of called functions).
